Fix coverity issues (#196)
* Fix static analysis issue (coverity)

* Improper use of negative value

* Uninitialized scalar field

* Logically dead code

* Uninitialized pointer field

+++ shell/platform/tizen/channels/text_input_channel.cc
@@ -272,7 +272,8 @@ void TextInputChannel::HandleMethodCall(
       size_t cursor_offset = selection_base_value - composing_start;
 
       active_model_->SetComposingRange(
-          flutter::TextRange(composing_base_value, composing_extent_value),
+          flutter::TextRange(static_cast<size_t>(composing_base_value),
+                             static_cast<size_t>(composing_extent_value)),
           cursor_offset);
     }
     SendStateUpdate(*active_model_);
